+++ Microsoft.Azure.Cosmos/src/CosmosClient.cs
@@ -46,7 +46,7 @@ namespace Microsoft.Azure.Cosmos
     /// </example>
     /// <example>
     /// This example create a <see cref="CosmosClient"/>, <see cref="Database"/>, and a <see cref="Container"/>.
-    /// The CosmosClient is created with the AccountEndpoint, AccountKey and configured to use "East US 2" region.
+    /// The CosmosClient is created with the AccountEndpoint, AccountKey or ResourceToken and configured to use "East US 2" region.
     /// <code language="c#">
     /// <![CDATA[
     /// using Microsoft.Azure.Cosmos;
@@ -167,10 +167,10 @@ namespace Microsoft.Azure.Cosmos
         /// performance guide at <see href="https://docs.microsoft.com/azure/cosmos-db/performance-tips"/>.
         /// </summary>
         /// <param name="accountEndpoint">The cosmos service endpoint to use</param>
-        /// <param name="accountKey">The cosmos account key to use to create the client.</param>
+        /// <param name="authKeyOrResourceToken">The cosmos account key or resource token to use to create the client.</param>
         /// <param name="clientOptions">(Optional) client options</param>
         /// <example>
-        /// The CosmosClient is created with the AccountEndpoint, AccountKey and configured to use "East US 2" region.
+        /// The CosmosClient is created with the AccountEndpoint, AccountKey or ResourceToken and configured to use "East US 2" region.
         /// <code language="c#">
         /// <![CDATA[
         /// using Microsoft.Azure.Cosmos;
@@ -195,7 +195,7 @@ namespace Microsoft.Azure.Cosmos
         /// </remarks>
         public CosmosClient(
             string accountEndpoint,
-            string accountKey,
+            string authKeyOrResourceToken,
             CosmosClientOptions clientOptions = null)
         {
             if (accountEndpoint == null)
@@ -203,9 +203,9 @@ namespace Microsoft.Azure.Cosmos
                 throw new ArgumentNullException(nameof(accountEndpoint));
             }
 
-            if (accountKey == null)
+            if (authKeyOrResourceToken == null)
             {
-                throw new ArgumentNullException(nameof(accountKey));
+                throw new ArgumentNullException(nameof(authKeyOrResourceToken));
             }
 
             if (clientOptions == null)
@@ -214,7 +214,7 @@ namespace Microsoft.Azure.Cosmos
             }
 
             this.Endpoint = new Uri(accountEndpoint);
-            this.AccountKey = accountKey;
+            this.AccountKey = authKeyOrResourceToken;
             CosmosClientOptions clientOptionsClone = clientOptions.Clone();
 
             DocumentClient documentClient = new DocumentClient(
@@ -238,7 +238,7 @@ namespace Microsoft.Azure.Cosmos
         /// </summary>
         internal CosmosClient(
             string accountEndpoint,
-            string accountKey,
+            string authKeyOrResourceToken,
             CosmosClientOptions cosmosClientOptions,
             DocumentClient documentClient)
         {
@@ -247,9 +247,9 @@ namespace Microsoft.Azure.Cosmos
                 throw new ArgumentNullException(nameof(accountEndpoint));
             }
 
-            if (accountKey == null)
+            if (authKeyOrResourceToken == null)
             {
-                throw new ArgumentNullException(nameof(accountKey));
+                throw new ArgumentNullException(nameof(authKeyOrResourceToken));
             }
 
             if (cosmosClientOptions == null)
@@ -263,7 +263,7 @@ namespace Microsoft.Azure.Cosmos
             }
 
             this.Endpoint = new Uri(accountEndpoint);
-            this.AccountKey = accountKey;
+            this.AccountKey = authKeyOrResourceToken;
 
             this.Init(cosmosClientOptions, documentClient);
         }
@@ -283,7 +283,7 @@ namespace Microsoft.Azure.Cosmos
         public virtual Uri Endpoint { get; }
 
         /// <summary>
-        /// Gets the AuthKey used by the client from the Azure Cosmos DB service.
+        /// Gets the AuthKey or resource token used by the client from the Azure Cosmos DB service.
         /// </summary>
         /// <value>
         /// The AuthKey used by the client.